프로필
유저 프로필 제작
닉네임 입력 중 단축키 작동 됨, 수정 필요
