+++ Assets/02.Script/UI/Backend/Ranking/RankingFrame.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 02c0f23a1d123c443b1b2b1d11b7ca68
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
