+++ Assets/02.Script/UI/Backend/Ranking/RankingViewer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0971983749d8d494d9373ac7924acc0a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
