+++ Assets/02.Script/UI/Backend/SingleTextPanel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3962370f358a9de44b76a34db7715625
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
