+++ Assets/02.Script/UI/Backend/UserInfo.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6c72f1e189bd58f4f901975e69b5e856
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
