+++ Assets/02.Script/UI/Backend/UserInfo/UserInfoViewController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c43b1bbdad50f844d98569b8a846a98c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
