--- Assets/02.Script/UI/Ranking.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9b16c91c62cf69c48a5ef38679df810b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
